integración semifront y formulas
